fixed the 'tag' type
git-svn-id: svn://orca.physics.unc.edu/Orca64@11968 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Experiments/MJDTestCryoVac/ORMJDTestCryostats/ORMJDTestCryoView.h
@@ -18,7 +18,7 @@
 //-------------------------------------------------------------
 #import "ORMJDPumpCartView.h"
 @interface ORMJDTestCryoView : ORMJDPumpCartView {
-	int tag;
+	NSUInteger tag;
 }
 - (NSUInteger) tag;
 - (void) setTag:(NSUInteger)aValue;
